+++ src/oldas/__main__.py
@@ -17,21 +17,8 @@ async def main() -> None:
             session, i="tag:google.com,2005:reader/item/6952dab95f45b77afe000dbf"
         )
     )
-    print(article.categories)
-    print(await article.mark_read(session))
-    article = await anext(
-        Articles.stream(
-            session, i="tag:google.com,2005:reader/item/6952dab95f45b77afe000dbf"
-        )
-    )
-    print(article.categories)
-    print(await article.mark_unread(session))
-    article = await anext(
-        Articles.stream(
-            session, i="tag:google.com,2005:reader/item/6952dab95f45b77afe000dbf"
-        )
-    )
-    print(article.categories)
+    for alternate in article.alternate:
+        print(alternate)
 
 
 if __name__ == "__main__":