+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # OldAS ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added support for loading up the alternates for an article.
+
 ## v0.1.0
 
 **Released: 2026-01-01**